fix(tasks): the board shows the command, and shows it once
Two things in the drawer.

**It said "Bash" where the thread said "Bash(cd ../goterm-workspace)".** Same
agent, same second, two different answers to "what is it doing" — and the less
useful one was on the screen built for watching work. The label was written in
internal/bot, so only the chat lane could reach it; the task runner passed the
bare tool name because that was all it had. The label now lives in internal/chat,
which both lanes already import, and its tests moved with it: a test living in
one lane's package could only ever check that lane.

**And the live row was rendered twice.** I inserted the block twice while
patching the file and did not look at the result. One now.
