+++ internal/taskrunner/runner.go
@@ -403,8 +403,12 @@ func (r *Runner) execute(ctx context.Context, task *coord.Task) {
 			progress.Text(chunk)
 		},
 		OnToolCall: func(name, input string) {
-			sess.NoteTool(name)
-			progress.Tool(name)
+			// The label, not the bare name: the board is the screen built for
+			// watching work, and "Bash" tells a watcher less than
+			// "Bash(cd ../goterm-workspace)" does.
+			label := chat.ToolLabel(name, input)
+			sess.NoteTool(label)
+			progress.Tool(label)
 			if name == "TodoWrite" && hasPendingTodos(input) {
 				todoPending = true
 			}
